Fix existing devfile tests for docker exec
