+++ pkg/devfile/adapters/docker/utils/utils_test.go
@@ -466,7 +466,7 @@ func TestAddProjectVolumeToComp(t *testing.T) {
 		hostConfig := container.HostConfig{
 			Mounts: tt.mounts,
 		}
-		AddProjectVolumeToComp(projectVolumeName, &hostConfig)
+		AddVolumeToComp(projectVolumeName, lclient.OdoSourceVolumeMount, &hostConfig)
 		if !reflect.DeepEqual(tt.want, hostConfig) {
 			t.Errorf("expected %v, actual %v", tt.want, hostConfig)
 		}
